Add files via upload
I created an expense tracker in Python. The expenses are displayed in a Pandas dataframe. The user is able to input their various expenses: food, transportation, utility, etc. With this tracker, the user can track how much they have spent of their monthly income and how much is left of the income.